feat: integrate runtime, SQL media, and dashboard updates (#19)
* feat: integrate runtime, SQL media, and dashboard updates


* fix: require a secure Pillow baseline


* fix: require a secure pypdf baseline


* fix: verify fsspec license provenance


* docs: refresh evaluation translations


* test: validate reviewed license overrides


---------

+++ src/ray/klein/runtime/graph/serve_rewriter.py
@@ -39,7 +39,7 @@ class _ProxyConfig:
     num_cpus: float
     concurrency: int | tuple[int, int]
     batch_size: int
-    batch_timeout: int
+    batch_timeout: float
     async_buffer_size: int
